+++ react/ImageList.tsx
@@ -12,6 +12,7 @@ export interface ImageListProps {
   height?: number
   preload?: boolean
   experimentalPreventLayoutShift?: boolean
+  experimentalSetExplicitDimensions?: boolean
 }
 
 function ImageList({
@@ -20,6 +21,7 @@ function ImageList({
   children,
   preload,
   experimentalPreventLayoutShift,
+  experimentalSetExplicitDimensions,
 }: PropsWithChildren<ImageListProps>) {
   const list = useListContext()?.list ?? []
   const { isMobile } = useDevice()
@@ -29,7 +31,8 @@ function ImageList({
     isMobile,
     height,
     preload,
-    experimentalPreventLayoutShift
+    experimentalPreventLayoutShift,
+    experimentalSetExplicitDimensions
   )
 
   const newListContextValue = list.concat(imageListContent)